Computer-Vision-#2 Implement the my_imfilter.m function.
Implemented my_imfilter.m to take an image and a filter as parameters
then created a padded version of the image using the 'symmetric' padding
method. After the image has been padded, the function filters the
image (once for single channel) and three times (one for each channel
in three channel images).