Computer-Vision-#5 Create and Implement my_edgePrewitt.m
Created and implemented my_edgePrewitt.m. This implementation uses the
Prewitt operators [-1 -1 -1; 0 0 0; 1 1 1] and [-1 0 1; -1 0 1; -1 0 1]
however there are others that we may try here. The function creates a
magnitude images where magnitudes kept are those exceeding the threshold
value taken as a parameter. The function also creates an orientation
image where the values are the direction of the gradient at that pixel.