Computer-Vision-#7 Create base for Marr-Hildreth algorithm.
Created the basic class for the Marr-Hildreth operator. Currently,
The Gaussian filter is computed manually by the algorithm. The function
computes the filter size based on the input parameter sigma and creates
a Gaussian filter of that size. The function uses the Gaussian filter
to blur the image and then computes the second order derivative of the
blurred image.