Computer-Vision-#7 Fix zero crossings.
Fixed the zero crossings portion of the algorithm. In its current state,
the edgeMagnitude image is created as all white (1) and then for
every pixel in the second derivative (Laplacian of the Gaussian in this
case) if the neighbor to the left, right, top, or bottom is black then
the pixel becomes black. This change signifies where the
second order derivative changed from positive to negative or made a
zero crossing.