Computer-Vision-#19 Add non-max suppression, and clean up code.
- Modified function description so that get_interest_points.m returns
a matrix of interest points as well as an array of locations of those
points.

- Capitalized constant alpha for good practices.

- Added grayscale conversion to the image received as a parameter.

- Added a second gaussian filter and modified their sizes. The purpose
of the second gaussian filter is to convolve the squared partials and
cell-wise product of the partials with the second filter.

- Added non-maximum suppression to the function to discard values that
are not a local maximum and those that are local maxima, but that do
not exceed a specified threshold.